refactor(matches): extract UpcomingMatch type to shared match-domain module
Move UpcomingMatch from home/UpcomingMatches/UpcomingMatches.tsx into a new
src/components/match/types.ts so match-domain components no longer have a
cross-feature dependency on the home domain.

- Create src/components/match/types.ts with the UpcomingMatch interface
- UpcomingMatches.tsx now imports and re-exports it (no breaking change)
- Update match.mapper.ts, MatchList, MatchesOverview, MatchesTabs,
  MatchesSlider (+ their stories) to import from @/components/match/types
- Export UpcomingMatch from the match barrel (src/components/match/index.ts)
