+++ src/components/match/MatchList/MatchList.tsx
@@ -7,7 +7,7 @@
 
 import { cn } from "@/lib/utils/cn";
 import { MatchTeaser } from "../MatchTeaser/MatchTeaser";
-import type { UpcomingMatch } from "@/components/home/UpcomingMatches/UpcomingMatches";
+import type { UpcomingMatch } from "@/components/match/types";
 
 export interface MatchListProps {
   /** Matches to display */
